Remove typings and move dependencies to package
Remove typings because typescript now works with npm modules.
Now we do not need this anymore:
```
	typings install
```
Now it is:
```
	npm i --save-dev @types/modulo_name
```

We are moving dependencies from yeoman install to concat them in
package.json to be installed later.

Now tests will fail because tsc is throwing errors, we have to fix them.